Move `barrier<thread_scope_thread>` into its own file
